backend/requirements.txt
django
djangorestframework
django-countries
git+https://github.com/behave/behave
behave-django
django-filter
requests